#5 FileService refactoring
