+++ api/src/main/java/com/blackpet/dndfileupload/file/presentation/FileController.java
@@ -1,12 +1,12 @@
 package com.blackpet.dndfileupload.file.presentation;
 
+import com.blackpet.dndfileupload.file.application.FileService;
 import com.blackpet.dndfileupload.file.domain.AttachFile;
 import com.blackpet.dndfileupload.file.infrastructure.FileRepository;
 import com.blackpet.dndfileupload.file.infrastructure.dto.AttachFileResponse;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.tomcat.util.http.fileupload.IOUtils;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
@@ -17,8 +17,6 @@ import java.io.IOException;
 import java.io.InputStream;
 import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
-import java.nio.file.Path;
-import java.nio.file.Paths;
 import java.util.UUID;
 
 @RestController
@@ -26,23 +24,11 @@ import java.util.UUID;
 @Slf4j
 public class FileController {
 
-  @Value("${upload.directory}")
-  private String uploadDirectory;
-
-  private final FileRepository fileRepository;
+  private final FileService fileService;
 
   @PostMapping("/files/upload")
   public ResponseEntity<AttachFileResponse> uploadFile(@RequestParam("file") MultipartFile file) throws IOException {
-
-    Path targetDirectory = Paths.get(uploadDirectory);
-
-    AttachFile attachFile = new AttachFile(file.getOriginalFilename(), file.getContentType(), file.getSize(), targetDirectory);
-
-    // copy physical file
-    copyFileToPath(file.getInputStream(), attachFile);
-
-    // save to DB
-    fileRepository.save(attachFile);
+    AttachFile attachFile = fileService.saveFile(file);
 
     AttachFileResponse response = AttachFileResponse.builder()
             .id(attachFile.getId())
@@ -54,10 +40,7 @@ public class FileController {
 
   @GetMapping("/files/download/{id}")
   public void downloadFile(@PathVariable UUID id, HttpServletResponse response) throws IOException {
-    AttachFile file = fileRepository.findById(id)
-            .orElseThrow(() -> new RuntimeException("첨부되지 않은 파일입니다."));
-
-    file.verifyDownload();
+    AttachFile file = fileService.getAttachFile(id);
 
     InputStream inputStream = new FileInputStream(file.getFilePath().toFile());
 
@@ -69,30 +52,11 @@ public class FileController {
 
   @DeleteMapping("/files/{id}")
   public ResponseEntity<Boolean> deleteFile(@PathVariable UUID id) throws IOException {
-    AttachFile file = fileRepository.findById(id)
-            .orElseThrow(() -> new RuntimeException("존재하지 않는 파일입니다."));
+    boolean res = fileService.deleteFile(id);
 
-    fileRepository.delete(file);
-    Files.delete(file.getFilePath());
-
-    return ResponseEntity.ok(true);
+    return ResponseEntity.ok(res);
   }
 
-
-  private void copyFileToPath(InputStream inputStream, AttachFile file) {
-    Path target = file.getFilePath();
-
-    try (inputStream) {
-      // create non exists directories
-      if (Files.notExists(target.getParent())) {
-        Files.createDirectories(target.getParent());
-      }
-      Files.copy(inputStream, target);
-    } catch (IOException e) {
-      e.printStackTrace();
-      throw new RuntimeException(String.format(("파일 업로드가 실패했습니다. (%s)"), file.getName()));
-    }
-  }
 
   @GetMapping("/hello")
   public String hello() {
